test(solver): split relation cache agreement tests
